+++ src/Compute/Compute/AvailabilitySets/StartAzureAvailabilitySetMigrationCommand.cs
@@ -21,13 +21,17 @@ using Microsoft.Azure.Management.Compute.Models;
 
 namespace Microsoft.Azure.Commands.Compute
 {
-    [Cmdlet("Start", ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "AvailabilitySetMigration", SupportsShouldProcess = true)]
+    [Cmdlet("Start", ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "AvailabilitySetMigration", DefaultParameterSetName = ResourceGroupNameParameterSet, SupportsShouldProcess = true)]
     [OutputType(typeof(PSComputeLongRunningOperation), typeof(PSAzureOperationResponse))]
     public class StartAzureAvailabilitySetMigrationCommand : AvailabilitySetBaseCmdlet
     {
+        protected const string ResourceGroupNameParameterSet = "ResourceGroupNameParameterSet";
+        protected const string InputObjectParameterSet = "InputObjectParameterSet";
+
         [Parameter(
            Mandatory = true,
            Position = 0,
+           ParameterSetName = ResourceGroupNameParameterSet,
            ValueFromPipelineByPropertyName = true,
            HelpMessage = "The resource group name.")]
         [ResourceGroupCompleter]
@@ -38,6 +42,7 @@ namespace Microsoft.Azure.Commands.Compute
         [Parameter(
             Mandatory = true,
             Position = 1,
+            ParameterSetName = ResourceGroupNameParameterSet,
             ValueFromPipelineByPropertyName = true,
             HelpMessage = "The availability set name.")]
         [ResourceNameCompleter("Microsoft.Compute/availabilitySets", "ResourceGroupName")]
@@ -46,7 +51,23 @@ namespace Microsoft.Azure.Commands.Compute
 
         [Parameter(
             Mandatory = true,
+            Position = 0,
+            ParameterSetName = InputObjectParameterSet,
+            ValueFromPipeline = true,
+            HelpMessage = "The availability set object.")]
+        [ValidateNotNullOrEmpty]
+        public PSAvailabilitySet InputObject { get; set; }
+
+        [Parameter(
+            Mandatory = true,
             Position = 2,
+            ParameterSetName = ResourceGroupNameParameterSet,
+            ValueFromPipelineByPropertyName = true,
+            HelpMessage = "The resource ID of the Flexible Virtual Machine Scale Set to migrate to.")]
+        [Parameter(
+            Mandatory = true,
+            Position = 1,
+            ParameterSetName = InputObjectParameterSet,
             ValueFromPipelineByPropertyName = true,
             HelpMessage = "The resource ID of the Flexible Virtual Machine Scale Set to migrate to.")]
         [ValidateNotNullOrEmpty]
@@ -58,6 +79,12 @@ namespace Microsoft.Azure.Commands.Compute
         public override void ExecuteCmdlet()
         {
             base.ExecuteCmdlet();
+
+            if (this.ParameterSetName.Equals(InputObjectParameterSet))
+            {
+                this.ResourceGroupName = this.InputObject.ResourceGroupName;
+                this.Name = this.InputObject.Name;
+            }
 
             if (this.ShouldProcess(Name, "Start Availability Set Migration to Virtual Machine Scale Set"))
             {